Address review comments.
   - Move compile_with_dpex into the core.compiler module.
   - The target and typing contexts are now not set from dpex_target
     inside the SpirvKernel.compile. They are passed in as args.
   - Fix tests impacted by these changes.
